backend/pyproject.toml
[project]
name = "talentstreamai-api"
version = "0.1.0"
description = "TalentStreamAI FastAPI service"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn[standard]==0.34.0",
  "pydantic-settings==2.8.1",
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.3.0",
  "beautifulsoup4==4.13.0",
  "python-docx==1.1.2",
  "pypdf==5.1.0",
  "httpx==0.28.1",
  "python-multipart==0.0.20",
]

[tool.uv]
package = false
